refactor: 이미지 업로드 로직 분리
