+++ pages/my-page/components/ProfileEditForm/ProfileEditForm.module.scss
@@ -62,7 +62,7 @@
   .camera-icon {
     z-index: 5;
     position: absolute;
-    bottom: 0rem;
+    bottom: 0;
     right: -0.5rem;
   }
 }
